create `commands` submodule, add `__main__.py`
