+++ pyproject.toml
@@ -59,7 +59,7 @@ msgspec = ">=0.11.0"
 jinja2 = { version = ">=3.1.2", optional = true }
 mako = { version = ">=1.2.4", optional = true }
 click = {version = ">=8.1.3", optional = true}
-rich = {version = ">=12.6.0", optional = true}
+rich = {version = ">=13.0.1", optional = true}
 importlib-metadata = { version = ">=5.2.0", python = "<3.10" }
 fast-query-parsers = "*"
 
@@ -140,7 +140,7 @@ full = [
 ]
 
 [tool.poetry.scripts]
-starlite = {callable = "starlite:cli.main.cli", extras = ["cli"] }
+starlite = {callable = "starlite:__main__", extras = ["cli"] }
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
