+++ starlite/__main__.py
@@ -1,0 +1,3 @@
+from starlite.cli.main import starlite_group
+
+starlite_group()  # pylint: disable=E1120
